migrated chart component
