+++ src/app/components/detail-modal/detail-modal.component.html
@@ -12,20 +12,7 @@
     <div class="table-wrap">
       <app-break-up-table [dataSource]="dataMatTable"></app-break-up-table>
     </div>
-    <div class="graph-wrap">
-      <div class="graph-title lead">{{ graphs[currentGraph].title }}</div>
-      <div class="graph-subtitle lead">
-        Glide your cursor across the chart for supplementary details
-      </div>
-      <span (click)="slideGraph()" class="slide-wrapper slide-left">
-        <img src="assets/icons/slide-arrow-icon.svg" alt="slide" />
-      </span>
-      <span (click)="slideGraph()" class="slide-wrapper slide-right">
-        <img src="assets/icons/slide-arrow-icon.svg" alt="slide" />
-      </span>
-      <app-doughnut-graph #graph [data]="graphData"></app-doughnut-graph>
-      <div class="summary-panel"></div>
-    </div>
+  <app-graph-container></app-graph-container> 
   </div>
   <div class="footer-panel">
     <div class="data-panel">
