+++ src/app/components/detail-modal/detail-modal.component.scss
@@ -15,56 +15,9 @@
       width: 500px;
       overflow: auto;
     }
-    .graph-wrap {
-      position: relative;
-      flex-grow: 1;
-      display: flex;
-      flex-direction: column;
-      .graph-title {
-        margin-top: 12px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        height: 30px;
-        font-size: 1.4rem;
-      }
-      .graph-subtitle {
-        height: 20px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        opacity: 0.7;
-        font-size: 0.9rem;
-      }
-      app-doughnut-graph {
-        flex-grow: 1;
-        display: flex;
-        align-items: center;
-      }
-      .slide-wrapper {
-        position: absolute;
-        top: 40%;
-        img {
-          width: 50px;
-          filter: invert(1);
-          opacity: 0.1;
-        }
-        &:hover {
-          cursor: pointer;
-          img {
-            filter: $theme-filter;
-            opacity: 0.7;
-          }
-        }
-      }
-      .slide-left {
-        transform: rotateY(180deg);
-        left: 5%;
-      }
-      .slide-right {
-        right: 5%;
-      }
-    }
+   app-graph-container{
+    flex-grow: 1;
+   } 
   }
   .footer-panel {
     border-top: 1px solid $border-line;
@@ -132,7 +85,7 @@
         margin-bottom: 2px;
         box-shadow: none !important;
       }
-      .graph-wrap {
+      app-graph-container {
         display: none;
       }
     }
