+++ src/app/components/graph-container/graph-container.component.html
@@ -1,1 +1,14 @@
-<p>graph-container works!</p>
+<div class="graph-wrap">
+  <div class="graph-title lead">{{ graphs[currentGraph].title }}</div>
+  <div class="graph-subtitle lead">
+    Glide your cursor across the chart for supplementary details
+  </div>
+  <span (click)="slideGraph()" class="slide-wrapper slide-left">
+    <img src="assets/icons/slide-arrow-icon.svg" alt="slide" />
+  </span>
+  <span (click)="slideGraph()" class="slide-wrapper slide-right">
+    <img src="assets/icons/slide-arrow-icon.svg" alt="slide" />
+  </span>
+  <app-doughnut-graph #graph [data]="graphData"></app-doughnut-graph>
+  <div class="summary-panel"></div>
+</div>
