+++ src/app/components/graph-container/graph-container.component.scss
@@ -1,0 +1,51 @@
+@import "../../styles/palette.scss";
+.graph-wrap {
+  position: relative;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  .graph-title {
+    margin-top: 12px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 30px;
+    font-size: 1.4rem;
+  }
+  .graph-subtitle {
+    height: 20px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    opacity: 0.7;
+    font-size: 0.9rem;
+  }
+  app-doughnut-graph {
+    flex-grow: 1;
+    display: flex;
+    align-items: center;
+  }
+  .slide-wrapper {
+    position: absolute;
+    top: 40%;
+    img {
+      width: 50px;
+      filter: invert(1);
+      opacity: 0.1;
+    }
+    &:hover {
+      cursor: pointer;
+      img {
+        filter: $theme-filter;
+        opacity: 0.7;
+      }
+    }
+  }
+  .slide-left {
+    transform: rotateY(180deg);
+    left: 5%;
+  }
+  .slide-right {
+    right: 5%;
+  }
+}
